:art: Broke out navbar button and new tests
