+++ src/components/Navigation/Navbar.js
@@ -1,17 +1,12 @@
 import React from "react";
-import {
-  StyledNav,
-  StyledThemeButton,
-  StyledThemeButtonContainer,
-} from "./NavbarStyles";
+import NavbarButton from "./NavbarButton";
+import { StyledNav, StyledThemeButtonContainer } from "./NavbarStyles";
 
 const Navbar = ({ theme, handleThemeToggle }) => {
   return (
     <StyledNav data-testid="app-navbar">
       <StyledThemeButtonContainer>
-        <StyledThemeButton onClick={handleThemeToggle}>
-          {theme.mode === "dark" ? "Light Mode 🌕" : "Dark Mode 🌑"}
-        </StyledThemeButton>
+        <NavbarButton theme={theme} handleThemeToggle={handleThemeToggle} />
       </StyledThemeButtonContainer>
     </StyledNav>
   );
